kirana_web_app/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
numpy==1.17.2
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
pandas==0.24.2
Pillow==6.0.1
PyQt5==5.13.1
